Refactor step sequence functions and variables
This commit refactors the step sequence functions and variables in the main.cpp file. It updates the stepUpSequence() and stepDownSequence() functions to use a for loop and an array to handle the step sequence logic. It also renames the variables to be more descriptive and removes unused variables. This refactoring improves the readability and maintainability of the code.